Remove old debugging code
In the previous release, on an unknown exception the error_handler
printed the traceback directly to stdout. This was left in by mistake
during development.

Now as before an error message is logged to error, and the traceback
is logged to debug. The traceback can be seen by a user by passing
the flag --verbose twice.